ReFiCollectModule -> SwapCollectModule
